linux_devices: move gui -> model
