+++ cmake/in-files/config.h.in
@@ -32,7 +32,6 @@
 #undef HAVE_UNISTD_H
 #cmakedefine OCPN_USE_UDEV_PORTS
 #cmakedefine OCPN_USE_SYSFS_PORTS
-#cmakedefine HAVE_LIBUSB_10
 #cmakedefine HAVE_DIRENT_H
 #cmakedefine HAVE_LIBGEN_H
 #cmakedefine HAVE_LIBUDEV
