+++ gui/include/gui/usb_devices.h
@@ -26,7 +26,7 @@
 #define USB_DEVICES_H
 
 #ifdef HAVE_LIBUSB_10
-#include "linux_devices.h"
+#include "model/linux_devices.h"
 #else
 static bool inline is_dongle_permissions_wrong() { return false; }
 static bool inline is_device_permissions_ok(const char*) { return true; }
