+++ gui/src/udev_rule_mgr.cpp
@@ -35,11 +35,13 @@
 #include <wx/stattext.h>
 #include <wx/textctrl.h>
 
-#include "udev_rule_mgr.h"
-#include "linux_devices.h"
+#include "model/linux_devices.h"
 #include "model/logger.h"
-#include "gui_lib.h"
 #include "model/ocpn_utils.h"
+
+#include "gui_lib.h"
+#include "udev_rule_mgr.h"
+
 
 static bool hide_dongle_dialog;
 static bool hide_device_dialog;
