+++ model/CMakeLists.txt
@@ -207,6 +207,8 @@ endif ()
 if (LINUX)
   list(APPEND HDRS ${MODEL_HDR_DIR}/comm_drv_n2k_socketcan.h )
   list(APPEND SRC ${MODEL_SRC_DIR}/comm_drv_n2k_socketcan.cpp)
+  list(APPEND HDRS ${MODEL_HDR_DIR}/linux_devices.h)
+  list(APPEND SRC ${MODEL_SRC_DIR}/linux_devices.cpp)
 endif ()
 
 if (LINUX AND NOT QT_ANDROID)
@@ -289,6 +291,12 @@ if (UNIX AND NOT APPLE)
   if (GLIB_FOUND)
     target_link_libraries(_OCPN_MODEL INTERFACE glib::glib)
   endif ()
+endif ()
+
+find_package(LIBUSB)
+if (LIBUSB_FOUND)
+  target_link_libraries(_OCPN_MODEL INTERFACE ocpn::libusb)
+  add_compile_definitions(_OCPN_MODEL INTERFACE HAVE_LIBUSB_10)
 endif ()
 
 if (LINUX)
